--- python/MooseDiscussion/database/config.json
@@ -1,1 +1,0 @@
-{"dim": 384, "space": "ip", "ef": 10, "ef_construction": 200, "element_count": 1865, "max_elements": 10000}
